[feat] rest api 공통화 작업 및 화면 수정
