+++ src/pages/main/Main.js
@@ -1,9 +1,8 @@
-import BoardGroupList from "./boardGroupList/BoardGroupList";
+import BoardGroupList from "../board/BoardGroupList";
 
 const Main = () => {
     return (
         <>
-{/*        <h3>로그인이 완료되었습니다.</h3>*/}
         <BoardGroupList/>
         </>
     )
